+++ src/core/session.cpp
@@ -415,14 +415,14 @@ bool session::get_uppercase_column_names() const
     }
 }
 
-bool session::get_next_sequence_value(std::string const & sequence, long & value)
+bool session::get_next_sequence_value(std::string const & sequence, long long & value)
 {
     ensureConnected(backEnd_);
 
     return backEnd_->get_next_sequence_value(*this, sequence, value);
 }
 
-bool session::get_last_insert_id(std::string const & sequence, long & value)
+bool session::get_last_insert_id(std::string const & sequence, long long & value)
 {
     ensureConnected(backEnd_);
